PT-598: Publish event accessibility fix
